Update: convert plugin in typescript
