+++ packages/plugins/cloud/README.md
@@ -4,4 +4,3 @@ A simple plugin giving instructions on how to deploy your Strapi local project t
 
 - Upload your project to GitHub if not already done.
 - Authenticate & Deploy on Strapi Cloud.
-
