+++ packages/plugins/cloud/admin/src/components/Initializer/index.tsx
@@ -5,24 +5,20 @@
  */
 
 import { useEffect, useRef } from 'react';
-
-import PropTypes from 'prop-types';
-
 import pluginId from '../../pluginId';
 
-const Initializer = ({ setPlugin }) => {
-  const ref = useRef();
-  ref.current = setPlugin;
+type InitializerProps = {
+  setPlugin: (id: string) => void;
+};
+
+const Initializer = ({ setPlugin }: InitializerProps) => {
+  const ref = useRef(setPlugin);
 
   useEffect(() => {
     ref.current(pluginId);
   }, []);
 
   return null;
-};
-
-Initializer.propTypes = {
-  setPlugin: PropTypes.func.isRequired,
 };
 
 export default Initializer;
