+++ packages/plugins/cloud/admin/src/components/PluginIcon/index.tsx
