+++ packages/plugins/cloud/admin/src/utils/getTrad.ts
@@ -1,5 +1,5 @@
 import pluginId from '../pluginId';
 
-const getTrad = (id) => `${pluginId}.${id}`;
+const getTrad = (id: string) => `${pluginId}.${id}`;
 
 export default getTrad;
