--- packages/plugins/cloud/admin/src/utils/index.js
@@ -1,1 +1,0 @@
-export { default as getTrad } from './getTrad';
