--- packages/plugins/cloud/server/destroy.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = () => {};
