--- packages/plugins/cloud/server/services/index.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = {};
