+++ packages/plugins/cloud/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "extends": "@strapi/typescript-utils/tsconfigs/admin",
+
+  "compilerOptions": {
+    "target": "ESNext",
+    "strict": true
+  },
+
+  "include": ["admin", "custom.d.ts"],
+
+  "exclude": [
+    "node_modules/",
+    "dist/",
+
+    // Do not include server files in the server compilation
+    "server/",
+    // Do not include test files
+    "**/*.test.ts"
+  ]
+}
